! Deleting a custom field from the front CF menu now works properly. ! Attempting to delete a custom field from the menu also prompts like the delete button does in the edit screen. ! Minor formatting improvements for custom fields in the posting interface.

+++ sd_source/SimpleDesk-AdminCustomField.php
@@ -197,10 +197,19 @@ function shd_admin_custom_save()
 	checkSession('request');
 
 	// Deletifyingistuffithingi?
-	if (isset($_REQUEST['delete']))
+	if (isset($_REQUEST['delete']) && !empty($_REQUEST['field']))
 	{
+		$_REQUEST['field'] = (int) $_REQUEST['field'];
 		$smcFunc['db_query']('', '
 			DELETE FROM {db_prefix}helpdesk_custom_fields
+			WHERE id_field = {int:field}',
+			array(
+				'field' => $_REQUEST['field'],
+			)
+		);
+
+		$smcFunc['db_query']('', '
+			DELETE FROM {db_prefix}helpdesk_custom_fields_values
 			WHERE id_field = {int:field}',
 			array(
 				'field' => $_REQUEST['field'],
